+++ govtool/frontend/src/components/molecules/GovernanceActionCardHeader.tsx
@@ -3,10 +3,12 @@ import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined";
 
 import { Tooltip, Typography } from "@atoms";
 import { useTranslation } from "@hooks";
+import { getMetadataDataMissingStatusTranslation } from "@/utils";
+import { MetadataValidationStatus } from "@/models";
 
 type GovernanceActionCardHeaderProps = {
   title?: string;
-  isDataMissing: string | boolean;
+  isDataMissing: MetadataValidationStatus | boolean;
 };
 
 export const GovernanceActionCardHeader = ({
@@ -22,7 +24,6 @@ export const GovernanceActionCardHeader = ({
         alignItems: "center",
         mb: "20px",
         overflow: "hidden",
-
       }}
       data-testid="governance-action-card-header"
     >
@@ -38,11 +39,17 @@ export const GovernanceActionCardHeader = ({
           ...(isDataMissing && { color: "#9E2323" }),
         }}
       >
-        {isDataMissing || title}
+        {(isDataMissing !== false &&
+          getMetadataDataMissingStatusTranslation(
+            isDataMissing as MetadataValidationStatus,
+          )) ||
+          title}
       </Typography>
       {isDataMissing && typeof isDataMissing === "string" && (
         <Tooltip
-          heading={isDataMissing}
+          heading={getMetadataDataMissingStatusTranslation(
+            isDataMissing as MetadataValidationStatus,
+          )}
           paragraphOne={t("govActions.dataMissingTooltipExplanation")}
           placement="bottom-end"
           arrow